prueba/requirements.txt
psycopg2-binary
redis
pandas
requests
python-dotenv
tqdm
fastapi
uvicorn
openai>=1.43.0

